Fixed rust typing bugs

+++ rust/src/composition.rs
@@ -2,4 +2,4 @@
 // Should be unsigned to prevent values less than 0,
 // and so can be shifted right to divide.
 // PermT's max value is the highest permutation degree supported.
-pub type PermT = usize;
+pub type PermT = u32;
